move blend factor to CCRenderComponent
