+++ cocos2d/core/components/CCRenderComponent.js
@@ -27,7 +27,9 @@ const Component = require('./CCComponent');
 const defaultVertexFormat = require('../renderer/webgl/vertex-format');
 const renderEngine = require('../renderer/render-engine');
 const RenderFlow = require('../renderer/render-flow');
+const BlendFactor = require('../platform/CCMacro').BlendFactor;
 const RenderData = renderEngine.RenderData;
+const gfx = renderEngine.gfx;
 
 /**
  * !#en
@@ -45,6 +47,53 @@ let RenderComponent = cc.Class({
     editor: CC_EDITOR && {
         executeInEditMode: true,
         disallowMultiple: true
+    },
+
+    properties: {
+        _srcBlendFactor: BlendFactor.SRC_ALPHA,
+        _dstBlendFactor: BlendFactor.ONE_MINUS_SRC_ALPHA,
+
+          /**
+         * !#en specify the source Blend Factor, this will generate a custom material object, please pay attention to the memory cost.
+         * !#zh 指定原图的混合模式，这会克隆一个新的材质对象，注意这带来的
+         * @property srcBlendFactor
+         * @type {macro.BlendFactor}
+         * @example
+         * sprite.srcBlendFactor = cc.macro.BlendFactor.ONE;
+         */
+        srcBlendFactor: {
+            get: function() {
+                return this._srcBlendFactor;
+            },
+            set: function(value) {
+                this._srcBlendFactor = value;
+                this._updateBlendFunc(true);
+            },
+            animatable: false,
+            type:BlendFactor,
+            tooltip: CC_DEV && 'i18n:COMPONENT.sprite.src_blend_factor'
+        },
+
+        /**
+         * !#en specify the destination Blend Factor.
+         * !#zh 指定目标的混合模式
+         * @property dstBlendFactor
+         * @type {macro.BlendFactor}
+         * @example
+         * sprite.dstBlendFactor = cc.macro.BlendFactor.ONE;
+         */
+        dstBlendFactor: {
+            get: function() {
+                return this._dstBlendFactor;
+            },
+            set: function(value) {
+                this._dstBlendFactor = value;
+                this._updateBlendFunc(true);
+            },
+            animatable: false,
+            type: BlendFactor,
+            tooltip: CC_DEV && 'i18n:COMPONENT.sprite.dst_blend_factor'
+        },
     },
     
     ctor () {
@@ -140,9 +189,33 @@ let RenderComponent = cc.Class({
     },
 
     setMaterial (material) {
-        material.updateHash();
         this._material = material;
-    }
+
+        this._updateBlendFunc();
+        material.updateHash();
+    },
+        
+    _updateBlendFunc: function (updateHash) {
+        if (this.srcBlendFactor === gfx.BLEND_SRC_ALPHA && this.dstBlendFactor === gfx.BLEND_ONE_MINUS_SRC_ALPHA) {
+            return;
+        }
+
+        if (!this._material) {
+            return;
+        }
+
+        var pass = this._material._mainTech.passes[0];
+        pass.setBlend(
+            gfx.BLEND_FUNC_ADD,
+            this._srcBlendFactor, this._dstBlendFactor,
+            gfx.BLEND_FUNC_ADD,
+            this._srcBlendFactor, this._dstBlendFactor
+        );
+
+        if (updateHash) {
+            this._material.updateHash();
+        }
+    },
 });
 RenderComponent._assembler = null;
 RenderComponent._postAssembler = null;
